Revert "feat: fetching profiles from server"
