+++ app/page.tsx
@@ -9,7 +9,6 @@ const Home: React.FC = () => {
     <FiltersProvider>
       <DefaultHeader />
       <Filters />
-      {/* @ts-expect-error Server Component */}
       <ProfileList />
     </FiltersProvider>
   )
